renames program components
