+++ frontend/src/App.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import {
-  AdoptClassroom,
-  HipHop,
-  Mentorship,
+  FrisbeeProgram,
+  ChildrenCamps,
+  EducationProgram,
   Programs,
-  PublicSpeaking,
-  YouthLeader,
+  Balwadis,
+  Culture,
 } from './Components/pages/Programs';
 import About from './Components/pages/About';
 import Blog from './Components/pages/Blog';
