+++ frontend/src/Components/pages/Programs/index.js
@@ -1,16 +1,16 @@
-import AdoptClassroom from './AdoptClassroom';
-import HipHop from './HipHop';
-import Mentorship from './Mentoship';
+import FrisbeeProgram from './Frisbee';
+import ChildrenCamps from './ChildrenCamps';
+import EducationProgram from './EducationProgram';
 import Programs from './Programs';
-import PublicSpeaking from './PublicSpeaking';
-import YouthLeader from './YouthLeader';
+import Balwadis from './Balwadis';
+import Culture from './Culture';
 
 export {
-    AdoptClassroom,
-    HipHop,
-    Mentorship,
+    FrisbeeProgram,
+    ChildrenCamps,
+    EducationProgram,
     Programs,
-    PublicSpeaking,
-    YouthLeader,
+    Balwadis,
+    Culture,
 }
